% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{compose}
\alias{compose}
\title{Compose permutations}
\usage{
compose(perm1, perm2)
}
\arguments{
\item{perm1}{a permtuation or a collection of permutations}

\item{perm2}{a permtuation or a collection of permutations}
}
\value{
The composition of the permutations
}
\description{
This function composes two given permtuations. One of the arguments
can be a collection of permutations, but not both at the same time. In this
case, every permutation in the collection is composed with the
other argument
}
\examples{
compose(c(3,1,2,4), c(4,1,3,2))
}

