% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{cycle2str}
\alias{cycle2str}
\title{Friendly display the cycles}
\usage{
cycle2str(cy)
}
\arguments{
\item{cy}{a list with the set of cycles}
}
\description{
Given a list with the cycles of a permutation, displays them in the standard cycle notation
}
\examples{
cycle2str(perm2cycles(c(1,5,2,3,4)))
}

