% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{cycles2perm}
\alias{cycles2perm}
\title{Get the permutation given the cycles}
\usage{
cycles2perm(cycles)
}
\arguments{
\item{cycles}{a list with the set of disjoint cycles}
}
\value{
The permutation in vector notation
}
\description{
Get the permutation as a vector given the set of cycles in which it factorizes
}
\examples{
cycles2perm(perm2cycles(c(1,5,2,3,4)))
}

