% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\docType{data}
\name{data.apa}
\alias{data.apa}
\title{Sample of permutations APA}
\format{Each row is a permtuation}
\description{
A rda file containing a sample of permutations of the American Psychology Association
}

