% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerMallows.R
\docType{data}
\name{.CAYLEY.DISTANCE}
\alias{.CAYLEY.DISTANCE}
\title{PerMallows: A Package for Sampling and Estimation in Mallows Models}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.CAYLEY.DISTANCE
}
\description{
This package provides methods for sampling from and estimating parameters of Mallows models
based on different distance measures such as Cayley and Ulam.
}
\keyword{Cayley}
\keyword{LIS}
\keyword{PerMallows}
\keyword{Ulam}
\keyword{internal}
