% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{generate.aux.files}
\alias{generate.aux.files}
\title{Generates the files for Ulam}
\usage{
generate.aux.files(perm.length)
}
\arguments{
\item{perm.length}{number of items in the permutations}
}
\value{
Nothing. Only writes in the current folder the axiliary files
}
\description{
Generates files for Ulam which are aimed to accelelrate the processes of counting
the number of permutations at ech distance, sampling and learning IFF these operations
are going to be computted more than once
}
\examples{
generate.aux.files(4)
}

