% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{identity.permutation}
\alias{identity.permutation}
\title{Generate identity the permutation}
\usage{
identity.permutation(perm.length)
}
\arguments{
\item{perm.length}{number of items in the permutation}
}
\value{
The identity permutation of the specified number of items
}
\description{
This function generates the identity permutation of a given number of items
}
\examples{
identity.permutation(3)
identity.permutation(7)
}

