% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{insert}
\alias{insert}
\title{Insert operator}
\usage{
insert(perm, i, j)
}
\arguments{
\item{perm}{a permutation}

\item{i}{position of the permutation}

\item{j}{position of the permutation}
}
\value{
The permutation in the input in which th eoperation has been applied
}
\description{
Given a permutation and two positions i, j, move item in position i to position j
}
\examples{
insert(c(1,2,3,4,5),5,2)
insert(c(1,2,3,4,5),2,5)
}

