% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{inverse.perm}
\alias{inverse.perm}
\title{Generate inverse permutation}
\usage{
inverse.perm(perm)
}
\arguments{
\item{perm}{a permutation or matrix of permutations}
}
\value{
The inverse permutation.
If the input is a matrix,the matrix with the inverses
}
\description{
This function generates the inverse of a given permutation. If the input
is a matrix of permutations, invert all the permutations in the input.
}
\examples{
inverse.perm(c(1,2,3,4))
inverse.perm(c(2,3,4,1))
data <- matrix(c(1,2,3, 4,1,4,3,2,1,2,4,3), nrow = 3, ncol = 4, byrow = TRUE)
inverse.perm(data)
}

