% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inverse.permutation}
\alias{inverse.permutation}
\title{Generate inverse permutation}
\usage{
inverse.permutation(perm)
}
\arguments{
\item{perm}{a permutation or matrix of permutations}
}
\value{
The inverse permutation.
If the input is a matrix,the matrix with the inverses
}
\description{
This function generates the inverse of a given permutation. If the input
is a matrix of permutations, invert all the permutations in the input.
}
\examples{
inverse.permutation(c(1,2,3,4))
inverse.permutation(c(2,3,4,1))
sample <- matrix(c(1,2,3, 4,1,4,3,2,1,2,4,3), nrow = 3, ncol = 4, byrow = TRUE)
inverse.permutation(sample)
}

