% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{inversion}
\alias{inversion}
\title{Inversion operator}
\usage{
inversion(perm, i)
}
\arguments{
\item{perm}{a permutation}

\item{i}{position of the permutation}
}
\value{
The permutation in the input with an inversion at the specified position
}
\description{
Given a permutation and a position, swap positions i and i+1
}
\examples{
inversion(c(1,2,3,4,5),2)
}

