% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{is.permutation}
\alias{is.permutation}
\title{Check if its argument is a permutation}
\usage{
is.permutation(perm)
}
\arguments{
\item{perm}{a vector (or a bidimensional matrix)}
}
\value{
TRUE iff perm is a valid permutation (or a matrix of valid permutations)
}
\description{
This function tests if the given argument is a permutation of the first
n natural integers (excluding 0)
}
\examples{
is.permutation(c(3,1,2,4))
is.permutation(c(6,1,2,3))
is.permutation(matrix(c(1,2,3, 4,1,4,3,2,1,2,4,3), nrow = 3, ncol = 4, byrow = TRUE))
}

