% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{maxi.dist}
\alias{maxi.dist}
\title{Get the maximum value of the distance ebtween permutations}
\usage{
maxi.dist(perm.length, dist.name = "kendall")
}
\arguments{
\item{perm.length}{number of items in the permutations}

\item{dist.name}{optional. One of: kendall (default), cayley, hamming, ulam}
}
\value{
The maximum value for the distance between the permutations
}
\description{
Compute the maximum posible value for the distance between two given permutations.
The distance can be kendall, cayley, hamming and ulam
}
\examples{
maxi.dist(4,"cayley")
maxi.dist(10,"ulam")
maxi.dist(4)
}

