% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{perm2decomp}
\alias{perm2decomp}
\title{Get the decomposition vector}
\usage{
perm2decomp(perm, dist.name = "kendall")
}
\arguments{
\item{perm}{the permutation}

\item{dist.name}{optional the name of the distance. One of: kendall (default), cayley, hamming}
}
\value{
The distance decomposition vector of the given permutation and distance. For the Kendall distance is the inversion vector
}
\description{
Given a permutation and a distance name generate the decomposition vector
}
\examples{
perm2decomp(c(1,2,4,3,5), "kendall")
perm2decomp(c(1,2,4,3,5), "cayley")
perm2decomp(c(1,2,4,3,5), "hamming")
}

