% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{r.derangement}
\alias{r.derangement}
\title{Generate a collection of derangements}
\usage{
r.derangement(n, perm.length)
}
\arguments{
\item{n}{number of permutations in the sample}

\item{perm.length}{number of items in the permutations}
}
\value{
A sample of permutations with the specified length
and number of cycles
}
\description{
Given a number of permutations and the number of items in the permutations
generate a sample of permutations with the specified length
and no fixed point
}
\examples{
r.perms.cycles(1, 4, 2)
}

