% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{swap}
\alias{swap}
\title{Swap two items of a permutation}
\usage{
swap(perm, i, j)
}
\arguments{
\item{perm}{a permutation}

\item{i}{position of the permutation}

\item{j}{position of the permutation}
}
\value{
The permutation in the input in which the two speicfied items have been swapped
}
\description{
Given a permutation and two position, swap both positions
}
\examples{
swap(c(1,2,3,4,5),2,5)
}

