\name{count.derangements}
\alias{count.derangements}
\title{Count derangements}
\usage{
count.derangements(perm.length)
}
\arguments{
  \item{perm.length}{number of items in the permutations}
}
\value{
the number of derangements of the specified length
}
\description{
Given the number of items in the permutations, how many
permutations are there with no fixed point?
}
\examples{
count.derangements(4)
}

