% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PerMallows.R
\name{count.derangements}
\alias{count.derangements}
\title{Count derangements}
\usage{
count.derangements(perm.length)
}
\arguments{
\item{perm.length}{number of items in the permutations}
}
\value{
The number of derangements of the specified length
}
\description{
Given the number of items in the permutations,
how many permutations are there with no fixed point?
}
\examples{
count.derangements(4)
}

