% Generated by roxygen2 (4.0.0): do not edit by hand
\name{count.perms.cycles}
\alias{count.perms.cycles}
\title{Count permutations with a given number of cycles}
\usage{
count.perms.cycles(perm.length, num.cycles)
}
\arguments{
\item{perm.length}{number of items in the permutations}

\item{num.cycles}{number of cycles of the permutations to count}
}
\value{
The number of permutations with the specified number of cycles
}
\description{
Given the number of items in the permutations,
how many permutations are there with with the given number of cycles?
}
\examples{
count.perms.cycles(4,2)
count.perms.cycles(4,1)
}

