\name{permutation2cycles}
\alias{permutation2cycles}
\title{Decompose a permutation in a set of cycles}
\usage{
permutation2cycles(perm)
}
\arguments{
  \item{perm}{a permutation}
}
\value{
the permutation in the input in which the operation has
been applied
}
\description{
Factor a given a permutation in the set of independent
cycles
}
\examples{
permutation2cycles(c(1,5,2,3,4))
}

